pyproject.toml
[project]
name = "city-wallet-data"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "httpx>=0.28.0",
    "polars>=1.0.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 135

[tool.ruff.lint]
select = ["E", "F", "I", "B", "ANN", "RUF"]
ignore = ["RUF100", "B904", "ANN401"]